wip: merge parsingv2 with parsing
