+++ tests/please/BUILD
@@ -9,16 +9,6 @@ neovim_test(
 )
 
 neovim_test(
-    name = "parsingv2_spec",
-    test_file = "parsingv2_spec.lua",
-    deps = [
-        "//lua/please:cursor",
-        "//lua/please:parsingv2",
-        "//tests/utils:temptree",
-    ],
-)
-
-neovim_test(
     name = "query_spec",
     test_file = "query_spec.lua",
     deps = [
